Move time provider package
