+++ pkg/auth/dependency/userverify/provider.go
@@ -3,9 +3,9 @@ package userverify
 import (
 	gotime "time"
 
-	"github.com/skygeario/skygear-server/pkg/auth/dependency/time"
 	"github.com/skygeario/skygear-server/pkg/core/auth/authinfo"
 	"github.com/skygeario/skygear-server/pkg/core/config"
+	"github.com/skygeario/skygear-server/pkg/core/time"
 
 	"github.com/skygeario/skygear-server/pkg/core/skyerr"
